src/Omnispindle/requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==24.1.0
aiohttp==3.9.4
aiosignal==1.3.1
aiosqlite==0.20.0
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
argcomplete==3.2.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astroid==3.1.0
asttokens==2.4.1
async-lru==2.0.4
asyncua==1.1.5
atlassian-python-api==3.41.16
attrs==23.2.0
Automat==22.10.0
babel==2.16.0
backoff==2.2.1
balena-sdk @ git+https://github.com/balena-io/balena-sdk-python.git@d4671120a24c415df020dbea62c72543dab8199f
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
black==24.1.1
bleach==6.1.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.3.3
cbor2==5.6.3
certifi==2025.1.31
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
constantly==23.10.4
contourpy==1.2.1
courlan==1.3.2
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.4
datasets==3.2.0
dateparser==1.2.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.15
docutils==0.20.1
durationpy==0.9
dvtTestKit==0.2.9
env==0.1.0
evaluate==0.4.3
executing==2.0.1
fastapi==0.115.7
fastjsonschema==2.19.1
fastmcp==0.4.1
filelock==3.17.0
flake8==7.0.0
flatbuffers==25.1.21
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
google-api-core==2.24.0
google-auth==2.29.0
google-cloud-bigquery==3.20.1
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.66.0
grafana-client==4.2.2
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.62.1
gssapi==1.8.3
h11==0.14.0
hatchling==1.27.0
htmldate==1.9.3
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
imagesize==1.4.1
importlib_metadata==7.1.0
importlib_resources==6.5.2
incremental==22.10.0
iniconfig==2.0.0
instructor==1.1.0
intervaltree==3.1.0
invoke==2.2.0
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
isort==5.13.2
jaraco.classes==3.4.0
jedi==0.19.1
jh2==5.0.7
Jinja2==3.1.4
jira==3.8.0
jmespath==1.0.1
joblib==1.4.0
journal==0.4.0
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jusText==3.0.1
keyring==24.3.1
kiss-headers==2.4.3
kiwisolver==1.4.5
krb5==0.5.1
kubernetes==31.0.0
libcst==1.2.0
loguru==0.7.3
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
maturin==1.8.1
mccabe==0.7.0
mcp==1.3.0
mdurl==0.1.2
mistune==3.0.2
mmh3==5.0.1
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy==1.8.0
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.6.0
netmiko==4.3.0
networkx==3.4.2
nh3==0.2.15
niquests==3.12.2
nltk==3.9.1
notebook_shim==0.2.3
ntc_templates==4.1.0
numpy==1.26.4
oauthlib==3.2.2
oculustestkit @ file:///Users/d.edens/lab/automation_testing_platform/OculusTestKit
onnxruntime==1.20.1
openai==1.14.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
paho-mqtt==2.1.0
pandas==2.2.2
pandas-stubs==2.2.2.240603
pandocfilters==1.5.1
paramiko==3.4.0
parso==0.8.3
pathspec==0.12.1
patsy==0.5.6
pexpect==4.9.0
pgzero==1.2.1
pillow
pine-client==0.2.0
pipenv==2024.4.1
pkginfo==1.10.0
platformdirs==4.1.0
playwright==1.50.0
pluggy==1.3.0
portalocker==3.1.1
posthog==3.9.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.5
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pubnub==7.3.2
pure-eval==0.2.2
pyarrow==19.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pycryptodomex==3.19.1
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pyee==12.1.1
pyflakes==3.2.0
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.8.0
pylint==3.1.0
pymongo==4.11.1
PyNaCl==1.5.0
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyOpenSSL==25.0.0
pyparsing==3.1.1
pypdf==5.1.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyre-check==0.9.19
pyre-extensions==0.0.30
pyserial==3.5
pyshorteners==1.0.1
PySocks==1.7.1
pyspnego==0.10.2
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qh3==1.3.2
ratelimit==2.2.1
readme_renderer==43.0
redis==5.2.0
referencing==0.33.0
regex==2024.11.6
requests==2.32.3
requests-kerberos==0.14.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.17.1
rsa==4.9
rstcloth==0.6.0
# rust_ingest @ file:///Users/d.edens/lab/regressiontestkit/rust_ingest/target/wheels/rust_ingest-0.1.0-cp311-cp311-macosx_11_0_arm64.whl#sha256=029fd6cb4a42a0b3539399a787e9f8f886eac1395f2beeabecfb253b57209d9f
sacrebleu==2.5.1
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.11.4
scp==0.14.5
seaborn==0.13.1
selenium==4.21.0
semver==3.0.2
Send2Trash==1.8.2
sentence-transformers==3.3.1
service-identity==23.1.0
shell_gpt==1.4.0
shellingham==1.5.4
six==1.17.0
slack-sdk==3.26.1
slacker==0.11.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.2.6
sphinx-argparse==0.5.2
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-autoprogram==0.1.9
sphinxcontrib-confluencebuilder==2.4.0
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.0
python-jose[cryptography]==3.3.0
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.18.0
# -e git+ssh://git@bitbucket.org/elementalmachines/testmonitor-sdk.git@4bd6a129c3ab2dff2f6bdd6b69a163b6112863b6#egg=testmonitor_sdk&subdirectory=../../../../OculusTestKit/testmonitor-sdk
TestSlide==2.7.1
textfsm==1.1.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tld==0.13
tokenizers==0.21.0
tomlkit==0.12.4
torch==2.5.1
tornado==6.4
tqdm==4.67.1
trafilatura==2.0.0
traitlets==5.14.1
transformers==4.48.1
trio==0.25.0
trio-websocket==0.11.1
trove-classifiers==2025.1.15.22
tuna==0.5.11
twine==5.0.0
Twisted==23.10.0
typeguard==2.13.3
typer==0.15.2
types-python-dateutil==2.8.19.20240106
types-pytz==2023.3.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
urllib3==2.3.0
urllib3-future==2.12.909
uvicorn==0.34.0
uvloop==0.21.0
verlib2==0.2.0
virtualenv==20.29.1
wassima==1.2.1
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==0.54.0
websockets==14.2
wrapt==1.16.0
wsproto==1.2.0
xonsh==0.15.1
xxhash==3.5.0
yarl==1.9.4
zipp==3.17.0
zope.interface==6.1
